requirements.txt
torch
transformers
datasets
faiss-cpu
streamlit
fastapi
uvicorn
tensorboard
evaluate
